Add files via upload
Hangman is a game where you guess the consonants of the word, where the vowels are provided. The admin provides the word and the category, and the player guesses the consonants. Each wrong guess takes away 5 points, while each correct guess gives the player 5 points. The player loses if his or her points becomes 0. Score is determined by remaining points divided by number of guesses.